requirements.txt
Django==1.6.5
South==1.0.2
amqp==1.4.7
anyjson==0.3.3
billiard==3.3.0.21
celery==3.1.19
coverage==3.6
django-celery==3.1.17
django-common-helpers==0.7.0
django-cron==0.3.5
django-tastypie==0.12.2
kombu==3.0.29
python-dateutil==2.4.2
python-mimeparse==0.1.4
pytz==2015.7
redis==2.10.3
six==1.10.0
sqlparse==0.1.18
